% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{draft_homework}
\alias{draft_homework}
\title{Create a homework assignment (starter codes and instructions)}
\usage{
draft_homework(..., course_config = config_load("config.yml"),
  credentials = git2r::cred_token())
}
\arguments{
\item{...}{parameters to overwrite global options}

\item{course_config}{used to override course parameters}

\item{credentials}{credential to push}
}
\value{
list(dir, repo, remote)
}
\description{
Create a homework assignment (starter codes and instructions)
}
