% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_upstream_get}
\alias{git_upstream_get}
\title{A wrapper function for git2r::branch_get_upstream}
\usage{
git_upstream_get(branch = git_head())
}
\description{
A wrapper function for git2r::branch_get_upstream
}
