% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_upstream_set}
\alias{git_upstream_set}
\title{A wrapper function for git2r::branch_set_upstream}
\usage{
git_upstream_set(name, branch = git_head())
}
\description{
A wrapper function for git2r::branch_set_upstream
}
