% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repo.R
\name{github_api_create}
\alias{github_api_create}
\title{Create a GitHub repository}
\usage{
github_api_create(org = NULL)
}
\arguments{
\item{org}{A string. Return API to create an organization repository if specified.}
}
\value{
List consisting of path and HTTP method (POST).
}
\description{
Create a GitHub repository
}
\examples{
github_api_create()
github_api_create(org = "myorg")

}
\seealso{
\url{https://developer.github.com/v3/repos/#create}
}
