% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repo.R
\name{github_api_delete}
\alias{github_api_delete}
\title{API path and method to delete a repository}
\usage{
github_api_delete(owner, repo)
}
\arguments{
\item{owner}{GitHub login or organization name.}

\item{repo}{Repository name.}
}
\value{
List consisting of path and HTTP method (DELETE).
}
\description{
API path and method to delete a repository
}
\examples{
github_api_delete("kenjisato", "my-great-repo")

}
\seealso{
\url{https://developer.github.com/v3/repos/#delete}
}
