% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repo.R
\name{github_api_edit}
\alias{github_api_edit}
\title{API path and method to edit repository information}
\usage{
github_api_edit(owner, repo)
}
\arguments{
\item{owner}{GitHub login or organization name.}

\item{repo}{Repository name.}
}
\value{
List consisting of path and HTTP method (GET).
}
\description{
API path and method to edit repository information
}
\examples{
github_api_edit("kenjisato", "my-great-repo")

}
\seealso{
\url{https://developer.github.com/v3/repos/#edit}
}
