% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repo.R
\name{github_api_get}
\alias{github_api_get}
\title{API path and method to get repository information}
\usage{
github_api_get(owner, repo)
}
\arguments{
\item{owner}{GitHub login or organization name.}

\item{repo}{Repository name.}
}
\value{
List consisting of path and HTTP method (GET).
}
\description{
API path and method to get repository information
}
\examples{
github_api_get(owner = "me", repo = "myrepo")

}
\seealso{
\url{https://developer.github.com/v3/repos/#get}
}
