% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_repo.R
\name{github_api_list}
\alias{github_api_list}
\title{API path and method to list repositories.}
\usage{
github_api_list(..., username = NULL, org = NULL)
}
\arguments{
\item{...}{Not used.}

\item{username}{A string for GitHub user account.}

\item{org}{A string for GitHub organization account.}
}
\value{
List consisting of path and HTTP method.
}
\description{
API path and method to list repositories.
}
\examples{
github_api_list(username = "kenjisato")

}
\seealso{
\url{https://developer.github.com/v3/repos/#list-your-repositories}
    \url{https://developer.github.com/v3/repos/#list-user-repositories}
    \url{https://developer.github.com/v3/repos/#list-organization-repositories}
}
