% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{hw_init}
\alias{hw_init}
\title{Create an assignment (starter codes and instructions)}
\usage{
hw_init(name, template = NULL, ...)
}
\arguments{
\item{name}{Common name for local and remote repos.}

\item{template}{key for template. If unspecified or nonexistent,
the first template is used.}

\item{...}{parameters passed to \code{hw_repo_create()}}
}
\value{
Remote repository URL
}
\description{
This function creates assignment_dir/shortname directory.
Put the template into the folder. A new repository for the course
organization will be created with the specified name
}
\examples{
\dontrun{
hw_init("assignment01", "template1")
}
}
