% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{start_course}
\alias{start_course}
\title{Set up a course directory.}
\usage{
start_course(config_name = "config.yml")
}
\arguments{
\item{config_name}{Output file name.}
}
\description{
Set up a course directory.
}
\examples{
\dontrun{
start_course()
}
}
