% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{start_course}
\alias{start_course}
\title{Set up a course directory.}
\usage{
start_course(dir = ".")
}
\arguments{
\item{dir}{Path to the course directory.}
}
\description{
The default structure will be something like below:
}
\details{
course
├── assignments
│   ├── asgm01
│   └── asgm02
├── collected
│   ├── asgm01-alice
│   ├── asgm01-bob
│   ├── asgm02-alice
│   └── asgm02-bob
├── solutions
│   ├── asgm01
│   └── asgm02
└── templates
    └── blank
}
\examples{
\dontrun{
start_course(".")
}
}
